Cleanup and comments
