+++ pirouette.cabal
@@ -17,7 +17,6 @@ library
   exposed-modules:
       Language.Pirouette.Example
       Language.Pirouette.Example.IsUnity
-      Language.Pirouette.Example.Syntax
       Language.Pirouette.PlutusIR.Builtins
       Language.Pirouette.PlutusIR.SMT
       Language.Pirouette.PlutusIR.ToTerm
